Rename testing_tools.py to generate.py and refactor tests for generating sorting.
