+++ spikeinterface/core/tests/test_job_tools.py
@@ -1,4 +1,4 @@
-from spikeinterface.core.testing_tools import generate_recording
+from spikeinterface.core import generate_recording
 
 from spikeinterface.core.job_tools import divide_segment_into_chunks, ensure_n_jobs, ensure_chunk_size, \
     ChunkRecordingExecutor
