+++ spikeinterface/core/tests/test_npzsortingextractor.py
@@ -2,7 +2,7 @@ import pytest
 from pathlib import Path
 
 from spikeinterface.core import NpzSortingExtractor
-from spikeinterface.core.testing_tools import create_sorting_npz
+from spikeinterface.core import create_sorting_npz
 
 if hasattr(pytest, "global_test_folder"):
     cache_folder = pytest.global_test_folder / "core"
