+++ spikeinterface/core/tests/test_waveform_tools.py
@@ -5,7 +5,7 @@ import platform
 
 import numpy as np
 
-from spikeinterface.core.testing_tools import generate_recording, generate_sorting
+from spikeinterface.core import generate_recording, generate_sorting
 from spikeinterface.core.waveform_tools import extract_waveforms_to_buffers #allocate_waveforms_buffers, distribute_waveforms_to_buffers
