+++ spikeinterface/preprocessing/tests/test_normalize_scale.py
@@ -2,7 +2,7 @@ import pytest
 from pathlib import Path
 
 from spikeinterface import set_global_tmp_folder
-from spikeinterface.core.testing_tools import generate_recording
+from spikeinterface.core import generate_recording
 
 from spikeinterface.preprocessing import normalize_by_quantile, scale, center, zscore
